# packages/mtg-core/src/mtg_core/api/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MTG Core API",
    "description": "Magic: The Gathering card database and analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sets": {
      "get": {
        "tags": [
          "sets"
        ],
        "summary": "List Sets",
        "description": "List all sets with optional filters.",
        "operationId": "list_sets_sets_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by set name",
              "title": "Name"
            },
            "description": "Filter by set name"
          },
          {
            "name": "set_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by set type",
              "title": "Set Type"
            },
            "description": "Filter by set type"
          },
          {
            "name": "include_online_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include online-only sets",
              "default": true,
              "title": "Include Online Only"
            },
            "description": "Include online-only sets"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sets/{code}": {
      "get": {
        "tags": [
          "sets"
        ],
        "summary": "Get Set",
        "description": "Get detailed information about a specific set.",
        "operationId": "get_set_sets__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sets/{code}/cards": {
      "get": {
        "tags": [
          "sets"
        ],
        "summary": "Get Set Cards",
        "description": "Get cards in a set with pagination.",
        "operationId": "get_set_cards_sets__code__cards_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Cards per page",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Cards per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetCardsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sets/{code}/analysis": {
      "get": {
        "tags": [
          "sets"
        ],
        "summary": "Get Set Analysis",
        "description": "Get complete analysis of a set including value, mechanics, tribal themes, and composition.",
        "operationId": "get_set_analysis_sets__code__analysis_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardSummary": {
        "properties": {
          "uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "flavor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flavor Name"
          },
          "mana_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mana Cost"
          },
          "cmc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmc"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Color Identity"
          },
          "rarity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rarity"
          },
          "set_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Code"
          },
          "collector_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Number"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "power": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power"
          },
          "toughness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toughness"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "image_small": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Small"
          },
          "price_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Usd"
          },
          "purchase_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Link"
          },
          "owned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owned"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CardSummary",
        "description": "Card summary for search results."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MechanicInfoResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "card_count": {
            "type": "integer",
            "title": "Card Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "top_cards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Cards"
          }
        },
        "type": "object",
        "required": [
          "name",
          "card_count",
          "description",
          "top_cards"
        ],
        "title": "MechanicInfoResponse",
        "description": "Mechanic information for API response."
      },
...
      "PriceTierBreakdownResponse": {
        "properties": {
          "bulk": {
            "type": "integer",
            "title": "Bulk",
            "description": "Cards $0-1"
          },
          "playable": {
            "type": "integer",
            "title": "Playable",
            "description": "Cards $1-10"
          },
          "chase": {
            "type": "integer",
            "title": "Chase",
            "description": "Cards $10-50"
          },
          "premium": {
            "type": "integer",
            "title": "Premium",
            "description": "Cards $50+"
          }
        },
        "type": "object",
        "required": [
          "bulk",
          "playable",
          "chase",
          "premium"
        ],
        "title": "PriceTierBreakdownResponse",
        "description": "Price tier distribution for API response."
      },
...
      "RarityDistributionResponse": {
        "properties": {
          "mythic": {
            "type": "integer",
            "title": "Mythic"
          },
          "rare": {
            "type": "integer",
            "title": "Rare"
          },
          "uncommon": {
            "type": "integer",
            "title": "Uncommon"
          },
          "common": {
            "type": "integer",
            "title": "Common"
          },
          "special": {
            "type": "integer",
            "title": "Special"
          }
        },
        "type": "object",
        "required": [
          "mythic",
          "rare",
          "uncommon",
          "common",
          "special"
        ],
        "title": "RarityDistributionResponse",
        "description": "Rarity distribution for API response."
      },
...
      "SetAnalysisResponse": {
        "properties": {
          "set_code": {
            "type": "string",
            "title": "Set Code"
          },
          "set_name": {
            "type": "string",
            "title": "Set Name"
          },
          "total_cards": {
            "type": "integer",
            "title": "Total Cards"
          },
          "value_summary": {
            "$ref": "#/components/schemas/SetValueSummaryResponse"
          },
          "mechanics": {
            "items": {
              "$ref": "#/components/schemas/MechanicInfoResponse"
            },
            "type": "array",
            "title": "Mechanics"
          },
          "tribal_themes": {
            "items": {
              "$ref": "#/components/schemas/TribalThemeResponse"
            },
            "type": "array",
            "title": "Tribal Themes"
          },
          "type_distribution": {
            "$ref": "#/components/schemas/TypeDistributionResponse"
          },
          "rarity_distribution": {
            "$ref": "#/components/schemas/RarityDistributionResponse"
          }
        },
        "type": "object",
        "required": [
          "set_code",
          "set_name",
          "total_cards",
          "value_summary",
          "mechanics",
          "tribal_themes",
          "type_distribution",
          "rarity_distribution"
        ],
        "title": "SetAnalysisResponse",
        "description": "Complete set analysis response."
      },
      "SetCardsResponse": {
        "properties": {
          "set_code": {
            "type": "string",
            "title": "Set Code"
          },
          "set_name": {
            "type": "string",
            "title": "Set Name"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/CardSummary"
            },
            "type": "array",
            "title": "Cards"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "set_code",
          "set_name",
          "cards",
          "page",
          "page_size",
          "total_count"
        ],
        "title": "SetCardsResponse",
        "description": "Response for cards in a set."
      },
      "SetDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "block": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block"
          },
          "base_set_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Set Size"
          },
          "total_set_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Set Size"
          },
          "is_online_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Online Only"
          },
          "is_foil_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Foil Only"
          },
          "keyrune_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyrune Code"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "SetDetail",
        "description": "Detailed set information."
      },
      "SetSummary": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "SetSummary",
        "description": "Set summary for list results."
      },
      "SetValueSummaryResponse": {
        "properties": {
          "total_value": {
            "type": "number",
            "title": "Total Value"
          },
          "total_value_foil": {
            "type": "number",
            "title": "Total Value Foil"
          },
          "average_value": {
            "type": "number",
            "title": "Average Value"
          },
          "median_value": {
            "type": "number",
            "title": "Median Value"
          },
          "chase_card_count": {
            "type": "integer",
            "title": "Chase Card Count"
          },
          "top_cards": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ],
              "type": "array",
              "maxItems": 3,
              "minItems": 3
            },
            "type": "array",
            "title": "Top Cards"
          },
          "top5_concentration": {
            "type": "number",
            "title": "Top5 Concentration"
          },
          "price_tiers": {
            "$ref": "#/components/schemas/PriceTierBreakdownResponse"
          }
        },
        "type": "object",
        "required": [
          "total_value",
          "total_value_foil",
          "average_value",
          "median_value",
          "chase_card_count",
          "top_cards",
          "top5_concentration",
          "price_tiers"
        ],
        "title": "SetValueSummaryResponse",
        "description": "Price/value summary for API response."
      },
      "SetsResponse": {
        "properties": {
          "sets": {
            "items": {
              "$ref": "#/components/schemas/SetSummary"
            },
            "type": "array",
            "title": "Sets"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "sets",
          "count"
        ],
        "title": "SetsResponse",
        "description": "Sets list response."
      },
...
      "TribalThemeResponse": {
        "properties": {
          "creature_type": {
            "type": "string",
            "title": "Creature Type"
          },
          "card_count": {
            "type": "integer",
            "title": "Card Count"
          },
          "percentage": {
            "type": "number",
            "title": "Percentage"
          }
        },
        "type": "object",
        "required": [
          "creature_type",
          "card_count",
          "percentage"
        ],
        "title": "TribalThemeResponse",
        "description": "Tribal theme for API response."
      },
...
      "TypeDistributionResponse": {
        "properties": {
          "creatures": {
            "type": "integer",
            "title": "Creatures"
          },
          "instants": {
            "type": "integer",
            "title": "Instants"
          },
          "sorceries": {
            "type": "integer",
            "title": "Sorceries"
          },
          "enchantments": {
            "type": "integer",
            "title": "Enchantments"
          },
          "artifacts": {
            "type": "integer",
            "title": "Artifacts"
          },
          "lands": {
            "type": "integer",
            "title": "Lands"
          },
          "planeswalkers": {
            "type": "integer",
            "title": "Planeswalkers"
          },
          "battles": {
            "type": "integer",
            "title": "Battles"
          },
          "other": {
            "type": "integer",
            "title": "Other"
          }
        },
        "type": "object",
        "required": [
          "creatures",
          "instants",
          "sorceries",
          "enchantments",
          "artifacts",
          "lands",
          "planeswalkers",
          "battles",
          "other"
        ],
        "title": "TypeDistributionResponse",
        "description": "Card type distribution for API response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}